requirements.txt
# Bindings for the hermes protocol 
hermes-python>=0.2
rpi.gpio

